+++ src/components/Header/Header.jsx
@@ -1,12 +1,10 @@
 import React, { Suspense } from 'react';
 import { useSelector } from 'react-redux';
 import { Outlet } from 'react-router-dom';
-
-import Logo from './Logo/Logo';
-import UserLogo from './UserLogo/UserLogo';
-import UserAuth from './UserAuth/UserAuth';
+import Logo from './Logo';
+import UserAuth from './UserAuth';
+import UserLogo from './UserLogo';
 import { selectAuthAuthenticated } from '../../redux/auth/authSelectors';
-
 import { Conteiner, StyledHeader } from './Header.styled';
 
 const Header = () => {
@@ -15,7 +13,7 @@ const Header = () => {
   return (
     <Conteiner>
       <StyledHeader>
-        <Logo authenticated={authenticated} />
+        <Logo />
         {authenticated ? <UserLogo /> : <UserAuth />}
         <Suspense fallback={<div>Loading page...</div>}>
           <Outlet />
