+++ src/components/Welcome/WaterConsumptionTracer/WaterСonsumptionTracker.jsx
@@ -1,10 +1,9 @@
-import { Link } from 'react-router-dom';
-
+import { WaterConsTrackerStyled } from './WaterСonsTrackerStyled.styled';
 import calendarDays from '../../../assets/images/icons/calendar-days.svg';
 import presentationCharBar from '../../../assets/images/icons/presentation-chart-bar.svg';
 import wrenchScrewdriver from '../../../assets/images/icons/wrench-screwdriver.svg';
 
-import { WaterConsTrackerStyled } from './WaterСonsTrackerStyled.styled';
+import { Link } from 'react-router-dom';
 
 export const WaterConsumptionTracker = () => {
   return (
