+++ src/components/Welcome/Welcome.jsx
@@ -1,7 +1,6 @@
 import { WaterConsumptionTracker } from './WaterConsumptionTracer/WaterСonsumptionTracker.jsx';
-import { WhyDrinkWater } from './WhyDrinkWater/WhyDrinkWater.jsx';
-
 import { Conteiner, WelcomeStyled } from './WelcomeStyled.styled.js';
+import { WhyDrinkWater } from './WhyDrinkWater/WhyDrinkWater.jsx';
 
 const Welcome = () => {
   return (
