+++ src/pages/Main/Main.jsx
@@ -1,28 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
+import { StyledMainPageBackgroun } from './Main.styled';
 import { Loader, Welcome } from 'components';
-import { LoaderStyledMain, StyledMainPageBackgroun } from './Main.styled';
 
 const Main = () => {
-  const [showLoader, setShowLoader] = useState(true);
-
-  useEffect(() => {
-    const loadData = async () => {
-      setShowLoader(false);
-    };
-    loadData();
-  }, []);
-
   return (
     <StyledMainPageBackgroun>
-      <StyledMainPageBackgroun>
-        {showLoader ? (
-          <LoaderStyledMain>
-            <Loader />
-          </LoaderStyledMain>
-        ) : (
-          <Welcome />
-        )}
-      </StyledMainPageBackgroun>
+      <Loader />
+      <Welcome />
     </StyledMainPageBackgroun>
   );
 };
